starter/requirements.txt
alembic==1.4.2
aniso8601==6.0.0
appdirs==1.4.3
astroid==2.2.5
attrs==19.3.0
awscli==1.18.114
Babel==2.8.0
blis==0.2.4
botocore==1.17.37
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
cymem==2.0.3
distlib==0.3.0
docutils==0.15.2
ecdsa==0.13.2
filelock==3.0.12
Flask==1.0.2
Flask-Cors==3.0.8
flask-marshmallow==0.12.0
Flask-Migrate==2.5.3
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-Script==2.0.5
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
future==0.17.1
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
iniconfig==1.0.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.10.0
joblib==0.15.1
lazy-object-proxy==1.4.0
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.6.0
mathparse==0.1.2
mccabe==0.6.1
more-itertools==8.4.0
murmurhash==1.0.2
nltk==3.5
numpy==1.18.5
packaging==20.4
pbr==5.4.5
Pint==0.12
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
psycopg2==2.8.5
psycopg2-binary==2.8.2
py==1.9.0
pyasn1==0.4.8
pycryptodome==3.3.1
PyJWT==1.7.1
pylint==2.3.1
pymongo==3.10.1
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.7.5
python-editor==1.0.4
python-jose==3.2.0
python-jose-cryptodome==1.3.2
pytz==2019.1
PyYAML==5.1.2
regex==2020.6.8
requests==2.23.0
rsa==4.5
s3transfer==0.3.3
six==1.12.0
spacy==2.1.9
SQLAlchemy==1.3.3
srsly==1.0.2
stevedore==1.32.0
thinc==7.0.8
toml==0.10.1
tqdm==4.46.1
typed-ast==1.3.5
urllib3==1.25.9
virtualenv==20.0.20
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wasabi==0.6.0
Werkzeug==0.15.2
wrapt==1.11.1
WTForms==2.3.1
zipp==3.1.0
